+++ composeApp/src/commonMain/kotlin/database/entity/TaskEntity.kt
@@ -5,16 +5,16 @@ import androidx.room.PrimaryKey
 
 
 @Entity
- data class TaskEntity(
+data class TaskEntity(
 
-    @PrimaryKey(autoGenerate = true) val pk :Int=0,
-     val ticket_id : String,
-     val ticket_number: String,
-     val ticket_state: String,
-     val level: String,
-     val location: String,
-     val site: String,
-     val region: String,
-     val province: String,
-     val city: String,
+    @PrimaryKey(autoGenerate = true) val pk: Long = 0,
+    val ticket_id: String,
+    val ticket_number: String,
+    val ticket_state: String,
+    val level: String,
+    val location: String,
+    val site: String,
+    val region: String,
+    val province: String,
+    val city: String,
 )
